Changed tower speed from int to float

+++ Game/core/src/GameObjects/Tower.java
@@ -18,13 +18,13 @@ public class Tower extends GameObject implements Drawable
     private StateMachine stateMachine;
     private Sprite sprite;
     private float range;
-    private int fireSpeed;
+    private float fireSpeed;
     private int damage;
     private int cost;
 
     private Enemy target;
 
-    public Tower(Vector2 position, float rotation, Texture texture, float range, int fireSpeed, int damage, int cost) {
+    public Tower(Vector2 position, float rotation, Texture texture, float range, float fireSpeed, int damage, int cost) {
         super(position, rotation);
         this.sprite = new Sprite(texture);
         sprite.setOrigin(sprite.getWidth()/2,sprite.getHeight()/2);
@@ -88,7 +88,7 @@ public class Tower extends GameObject implements Drawable
         return sprite.getHeight();
     }
 
-    public int GetFireSpeed() {
+    public float GetFireSpeed() {
         return fireSpeed;
     }
